+++ test/dummy/config/initializers/recording_studio_wordpress_plugin_template.rb
@@ -1,5 +1,5 @@
 # frozen_string_literal: true
 
-GemTemplate.configure do |config|
+RecordingStudioWordpressPluginTemplate.configure do |config|
   # config.api_key = "..."
 end